requirements.txt
biopython>=1.79
numpy>=1.21.0